move multiaddress validation to primitives
