+++ storage-provider/server/src/config.rs
@@ -9,7 +9,7 @@ use clap::Args;
 use libp2p::{identity::Keypair, Multiaddr, PeerId};
 use polka_storage_provider_common::config::sealing::SealingConfiguration;
 use primitives::{
-    p2p::keypair_value_parser,
+    p2p::{keypair_value_parser, validate_tcp_multiaddr, validate_ws_multiaddr},
     proofs::{RegisteredPoStProof, RegisteredSealProof},
 };
 use serde::{de::Error, Deserialize, Deserializer};
@@ -40,42 +40,11 @@ fn default_node_address() -> Url {
 fn default_p2p_tcp_multiaddr() -> Multiaddr {
     Multiaddr::from_str("/ip4/127.0.0.1/tcp/8002").expect("value should be a valid Multiaddr")
 }
-fn validate_tcp_multiaddr(s: &str) -> Result<Multiaddr, String> {
-    const IP4_TCP: [&str; 2] = ["ip4", "tcp"];
-    const IP6_TCP: [&str; 2] = ["ip6", "tcp"];
-    let multiaddress = Multiaddr::from_str(s).map_err(|err| err.to_string())?;
-    let protocols = multiaddress.protocol_stack().collect::<Vec<_>>();
-    if protocols.is_empty() {
-        // Not sure if this is even possible, but checking doesn't hurt
-        return Err(format!("No protocols were detected for {s}"));
-    }
-    // ip6 isn't tested but just like above, it doesn't hurt to check
-    if protocols.as_slice() != IP4_TCP && protocols.as_slice() != IP6_TCP {
-        return Err(format!("Unsupported protocol stack: {:?}", protocols));
-    }
-    Ok(multiaddress)
-}
 
 fn default_p2p_ws_multiaddr() -> Multiaddr {
     Multiaddr::from_str("/ip4/127.0.0.1/tcp/8003/ws").expect("value should be a valid Multiaddr")
 }
 
-fn validate_ws_multiaddr(s: &str) -> Result<Multiaddr, String> {
-    const IP4_TCP_WS: [&str; 3] = ["ip4", "tcp", "ws"];
-    const IP6_TCP_WS: [&str; 3] = ["ip6", "tcp", "ws"];
-
-    let multiaddress = Multiaddr::from_str(s).map_err(|err| err.to_string())?;
-    let protocols = multiaddress.protocol_stack().collect::<Vec<_>>();
-    if protocols.is_empty() {
-        // Not sure if this is even possible, but checking doesn't hurt
-        return Err(format!("No protocols were detected for {s}"));
-    }
-    // ip6 isn't tested but just like above, it doesn't hurt to check
-    if protocols.as_slice() != IP4_TCP_WS && protocols.as_slice() != IP6_TCP_WS {
-        return Err(format!("Unsupported protocol stack: {:?}", protocols));
-    }
-    Ok(multiaddress)
-}
 #[derive(Debug, Clone, Deserialize, Args)]
 #[group(multiple = true, conflicts_with = "config")]
 #[serde(deny_unknown_fields)]
